% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{carbon_collect}
\alias{carbon_collect}
\title{Collect step logs from a pipeline object}
\usage{
carbon_collect(x)
}
\arguments{
\item{x}{An object returned by `carbon_step()`.}
}
\value{
A tibble (possibly empty).
}
\description{
`carbon_step()` stores logs as an attribute on the returned object.
}
