% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{carbon_default_output_dir}
\alias{carbon_default_output_dir}
\title{Default output directory registered by carbon_init()}
\usage{
carbon_default_output_dir()
}
\value{
The registered output directory, or NULL.
}
\description{
Default output directory registered by carbon_init()
}
\keyword{internal}
