% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{carbon_default_output_file}
\alias{carbon_default_output_file}
\title{Default output file registered by carbon_init()}
\usage{
carbon_default_output_file()
}
\value{
The registered output file name, or "emissions_r.csv".
}
\description{
Default output file registered by carbon_init()
}
\keyword{internal}
