% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{carbon_default_tracker}
\alias{carbon_default_tracker}
\title{Get the default tracker registered by carbon_init()}
\usage{
carbon_default_tracker()
}
\value{
The CodeCarbon tracker registered by the most recent
  [carbon_init()] call.
}
\description{
Get the default tracker registered by carbon_init()
}
\keyword{internal}
