% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycarbon_core.R
\name{carbon_track_all}
\alias{carbon_track_all}
\title{Track Emissions for a list of functions}
\usage{
carbon_track_all(tasks, tracker)
}
\arguments{
\item{tasks}{list of functions and arguments like list(fun = function, args = list(a = 1, b = 2))}

\item{tracker}{codecarbon EmissionsTracker object}
}
\value{
A tibble of function results and emissions
}
\description{
Track Emissions for a list of functions
}
