% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_view.R
\name{carbon_view}
\alias{carbon_view}
\title{Launch the carbon_view shiny app}
\usage{
carbon_view(use_browser = TRUE)
}
\description{
\code{carbon_view} Launches the app to analyze code carbon data with multiples techniques.
}
