% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycarbon_core.R
\name{tracker_stop}
\alias{tracker_stop}
\title{Stop Carbon Tracker}
\usage{
tracker_stop(tracker)
}
\arguments{
\item{tracker}{codecarbon EmissionsTracker object}
}
\value{
Stops Global Emissions Tracker
}
\description{
Stop Carbon Tracker
}
