% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycarbon_core.R
\name{carbon_init}
\alias{carbon_init}
\title{Initialize Carbon Tracker}
\usage{
carbon_init(
  project_name = "rtest",
  measure_power_secs = 10,
  tracking_mode = "machine"
)
}
\arguments{
\item{project_name}{Name of the project}

\item{measure_power_secs}{Seconds between power measurements}

\item{tracking_mode}{Tracking mode (default "machine")}
}
\value{
A codecarbon EmissionsTracker object
}
\description{
Initialize Carbon Tracker
}
