% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycarbon_core.R
\name{carbon_track}
\alias{carbon_track}
\title{Track emissions for a single function}
\usage{
carbon_track(fun, ..., tracker, task_id = uuid::UUIDgenerate())
}
\arguments{
\item{fun}{Arbitrary function}

\item{...}{Function inputs}

\item{tracker}{codecarbon EmissionsTracker object}

\item{task_id}{Unique ID of function task}
}
\value{
A tibble of function results and emissions
}
\description{
Track emissions for a single function
}
