% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycarbon_core.R
\name{tracker_start}
\alias{tracker_start}
\title{Start Carbon Tracker}
\usage{
tracker_start(tracker)
}
\arguments{
\item{tracker}{codecarbon EmissionsTracker object}
}
\value{
Starts Global Emissions Tracker
}
\description{
Start Carbon Tracker
}
